requirements.txt
# 小红书爬虫项目依赖
requests>=2.31.0
python-dotenv>=1.2.1
psycopg2-binary>=2.9.11
# 其他依赖都是Python标准库，不需要额外安装
